Extract service session model
