+++ session/dto.go
@@ -17,67 +17,5 @@
 
 package session
 
-import (
-	"time"
-
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/gofrs/uuid"
-	"github.com/mysteriumnetwork/node/identity"
-	"github.com/mysteriumnetwork/node/market"
-	"github.com/mysteriumnetwork/node/session/event"
-)
-
 // ID represents session id type.
 type ID string
-
-// PaymentEngine is responsible for interacting with the consumer in regard to payments.
-type PaymentEngine interface {
-	Start() error
-	WaitFirstInvoice(time.Duration) error
-	Stop()
-}
-
-// Session structure holds all required information about current session between service consumer and provider.
-type Session struct {
-	ID           ID
-	ConsumerID   identity.Identity
-	AccountantID common.Address
-	Proposal     market.ServiceProposal
-	ServiceID    string
-	CreatedAt    time.Time
-	done         chan struct{}
-}
-
-// Done returns readonly done channel.
-func (s *Session) Done() <-chan struct{} {
-	return s.done
-}
-
-func (s Session) toEvent(status event.Status) event.AppEventSession {
-	return event.AppEventSession{
-		Status: status,
-		Service: event.ServiceContext{
-			ID: s.ServiceID,
-		},
-		Session: event.SessionContext{
-			ID:           string(s.ID),
-			StartedAt:    s.CreatedAt,
-			ConsumerID:   s.ConsumerID,
-			AccountantID: s.AccountantID,
-			Proposal:     s.Proposal,
-		},
-	}
-}
-
-// NewSession creates a blank new session with an ID.
-func NewSession() (*Session, error) {
-	uid, err := uuid.NewV4()
-	if err != nil {
-		return nil, err
-	}
-	return &Session{ID: ID(uid.String())}, nil
-}
-
-// ServiceConfiguration defines service configuration from underlying transport mechanism to be passed to remote party
-// should be serializable to json format.
-type ServiceConfiguration interface{}
